move types out into their own file
