+++ src/components/Table/Next/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./TableTypes";
